Move responses to JSON files and handle exceptions
Moved response choices to dedicated JSON files and enabled multipart
responses, which are composed by simple concatenation with spaces in
between.

Also, implemented better exception handling. Now requests to LUIS return
results, i.e. Async<Result<string, string>>, which means that if the
request fails, an exception will not be thrown, but instead a Bad result
will be returned.

In addition, implemented an API for C# to handle Results. They are
converted to a value, or an exception is thrown if the result is Bad.
This is a more conventional OO approach and spares the pains of pattern
matching in C#.